requirements.txt
streamlit
requests
pandas
plotly.express
openpyxl
